+++ design-system/packages/editor/src/index.ts
@@ -2,4 +2,4 @@
 
 export { EditorPopover } from './EditorPopover';
 
-export type { EditorPopoverProps, EditorPopoverRef } from './types';
+export type { EditorPopoverProps, EditorPopoverRef } from './EditorPopover';
